LGA-2604: ✅  Implement PR reviews and done some refactoring
