+++ behave/features/cla_public/public_exit_page.feature
@@ -1,6 +1,6 @@
 @cla_public
 Feature:
-  - Applying for legal aid on the "Check if I can get legal aid" website
+  - Testing the "Exit this page" journey through all possibilities available for a user
 
   Background: Start page
     Given I have selected the start now button on the start page
@@ -12,8 +12,8 @@ Feature:
       | category       |
       | Domestic abuse |
     And I am taken to the "Choose the option that best describes your personal situation" page located on "/scope/diagnosis/n43n3"
-    Then I click on the "Exit this page" button
-    And I am diverted to the BBC website "https://www.bbc.co.uk/weather"
+    Then The "Exit this page" button is on the page and I click it
+    And I am diverted to the BBC website
 
   @exit-behaviour-tabbing
   Scenario: Describes how to exit from the domestic abuse section and the website safely when pressing the tab key 2 times
@@ -21,8 +21,8 @@ Feature:
       | category       |
       | Domestic abuse |
     Then I am taken to the "Choose the option that best describes your personal situation" page located on "/scope/diagnosis/n43n3"
-    Then I press the "tab" key "2 times" on the keyboard
-    And I am diverted to the BBC website "https://www.bbc.co.uk/weather"
+    Then I press the "tab" key 2 times on the keyboard
+    And I am diverted to the BBC website
 
   @exit-behaviour-press-shift
   Scenario: Describes how to exit from the domestic abuse section and the website safely when pressing the shift key 3 times
@@ -30,5 +30,5 @@ Feature:
       | category       |
       | Domestic abuse |
     Then I am taken to the "Choose the option that best describes your personal situation" page located on "/scope/diagnosis/n43n3"
-    Then I press the "shift" key "3 times" on the keyboard
-    And I am diverted to the BBC website "https://www.bbc.co.uk/weather"
+    Then I press the "shift" key 3 times on the keyboard
+    And I am diverted to the BBC website
